+++ types.ts
@@ -1,5 +1,5 @@
 
-export const APP_VERSION = '1.3.7';
+export const APP_VERSION = '1.3.8';
 
 export enum Role {
   USER = 'user',